pyproject.toml
[project]
name = "amplifier-module-provider-github-copilot"
version = "2.0.0"
description = "GitHub Copilot provider for Amplifier - Three-Medium Architecture"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Amplifier Team", email = "amplifier@microsoft.com"}
]
keywords = ["amplifier", "github-copilot", "llm", "provider", "ai", "language-model"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    # GitHub Copilot CLI SDK
    # - v0.2.0: Breaking changes - SubprocessConfig, send(prompt, attachments=...)
    "github-copilot-sdk>=0.2.0,<0.3.0",
    # PyYAML for config loading
    "pyyaml>=6.0",
    # Note: amplifier-core is provided by Amplifier runtime when module is loaded
    # Listed here for type hints only; not installed at runtime
]

[project.urls]
Homepage = "https://github.com/microsoft/amplifier-module-provider-github-copilot"
Repository = "https://github.com/microsoft/amplifier-module-provider-github-copilot"
Documentation = "https://github.com/microsoft/amplifier-module-provider-github-copilot#readme"

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pyright>=1.1.350",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=7.0.0",
    # amplifier-core from PyPI (binary wheels, no Rust compilation needed)
    "amplifier-core>=1.0.7",
]

[project.entry-points."amplifier.modules"]
provider-github-copilot = "amplifier_module_provider_github_copilot:mount"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["amplifier_module_provider_github_copilot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
include = ["amplifier_module_provider_github_copilot", "tests"]
exclude = ["run_*.py"]
# Use system Python where amplifier-core is installed
pythonPlatform = "Linux"
# Exclude stub-related errors for pytest (has inline types but pyright strict mode is pedantic)
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
# Live tests excluded by default (CI-friendly) - run with: pytest -m live
# Or run all tests with: pytest -m ""
# Contract: tests run and FAIL if credentials missing, never skip
addopts = "-m 'not live'"
markers = [
    "contract: marks tests as contract compliance (run with -m contract)",
    "pure: marks tests as pure/no-IO (run with -m pure)",
    "canary: marks tests as SDK canary (run with -m canary)",
    "live: marks tests requiring real credentials (run explicitly with -m live)",
    "sdk_assumption: marks tests verifying SDK types/shapes exist",
]

# Note: [dependency-groups] removed — use [project.optional-dependencies] dev instead
# The pytest-cov dependency has been consolidated into [project.optional-dependencies] dev
